backend/requirements.txt
Django==1.11.2
djangorestframework==3.6.3
django-cors-headers==2.1.0

# Generate test data
django-autofixture==0.12.1
djangorestframework-jwt==1.11.0

